+++ src/features/cart/cartSlice.test.js
@@ -1,12 +1,4 @@
-/* eslint-disable no-unused-vars */
-import {
-  ADD_CHOSEN_INGREDIENT,
-  ADD_CHOSEN_BUN,
-  REMOVE_CHOSEN_INGREDIENT,
-  ELEMENT_SORTED_BY_DND,
-  RESET_CHOSEN_INGREDIENTS,
-} from '../../utils/constants';
-import reducer from './cart';
+import reducer from './cartSlice';
 
 const state = {
   chosenIngredients: [],
@@ -21,7 +13,7 @@ describe('cart reducer', () => {
     expect(reducer(undefined, {})).toEqual(state);
   });
   it('should set chosenIngredients', () => {
-    expect(reducer(state, { type: ADD_CHOSEN_INGREDIENT, payload: mockIngredients }))
+    expect(reducer(state, { type: 'cart/addIngredient', payload: mockIngredients }))
       .toEqual(
         expect.objectContaining({
           chosenIngredients: expect.arrayContaining([...mockIngredients]),
@@ -29,7 +21,7 @@ describe('cart reducer', () => {
       );
   });
   it('should set chosenBun', () => {
-    expect(reducer(state, { type: ADD_CHOSEN_BUN, payload: mockBun }))
+    expect(reducer(state, { type: 'cart/addBun', payload: mockBun }))
       .toEqual(
         expect.objectContaining({
           chosenBun: expect.objectContaining({ bun: true }),
@@ -40,18 +32,18 @@ describe('cart reducer', () => {
     expect(reducer({
       chosenIngredients: mockIngredients,
       chosenBun: mockIngredients,
-    }, { type: RESET_CHOSEN_INGREDIENTS }))
+    }, { type: 'cart/resetIngredients' }))
       .toStrictEqual(state);
   });
   it('should remove ingredient', () => {
     expect(reducer({
       chosenIngredients: mockIngredients,
-    }, { type: REMOVE_CHOSEN_INGREDIENT, payload: { positionIndex: 2 } }))
+    }, { type: 'cart/removeIngredient', payload: { positionIndex: 2 } }))
       .toEqual({ chosenIngredients: expect.not.arrayContaining(['eded']) });
   });
   it('should sort ingredients', () => {
     expect(reducer({ chosenIngredients: mockIngredients }, {
-      type: ELEMENT_SORTED_BY_DND,
+      type: 'cart/sortIngredients',
       payload: { positionIndex: 3, targetIndex: 0 },
     }))
       .toStrictEqual({ chosenIngredients: [undefined, 1, 3, 'eded'] });
